+++ src/lib/formatRegisterFormValues.ts
@@ -9,20 +9,26 @@ function capitalizeFirst(value: string) {
 	return value.charAt(0).toUpperCase() + value.slice(1).toLowerCase();
 }
 
-function getFirsWord(value: string) {
-	return value.split(" ")[0];
+function getFirstName(value: string) {
+	if (!value) return "";
+
+	const split = value.split(" ");
+	if (split.length < 2) return "";
+
+	return capitalizeFirst(split[0]);
 }
+
 export default function formatRegisterFormValues(values: Values) {
 	const parseValues = {
 		...values,
-		gender: values.genderIdentity,
-		firstName: capitalizeFirst(getFirsWord(values.firstName)),
+		firstName: getFirstName(values.firstName),
 		neighborhood: capitalizeFirst(values.neighborhood),
 		dateOfBirth: getDateFormat(values.dateOfBirth),
-		acceptsOnlineSupport: "yes",
 	};
-	console.log(parseValues);
-	return JSON.stringify(parseValues)
+
+	const replaceYesNoWithBoolean = JSON.stringify(parseValues)
 		.replace(/"yes"/g, "true")
 		.replace(/"no"/g, "false");
+
+	return replaceYesNoWithBoolean;
 }
